pyproject.toml
[project]
name = "warry-about"
version = "0.1.0"
description = "お悩み相談チャットアプリのバックエンド"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Web Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "aiomysql>=0.2.0",
    "alembic>=1.14.0",
    # LLM
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    # Auth
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    # Validation & Settings
    "pydantic[email]>=2.0.0",
    "pydantic-settings>=2.0.0",
    # Utils
    "python-multipart>=0.0.12",
    "httpx>=0.27.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "factory-boy>=3.3.0",
    "mypy>=1.13.0",
    "ruff>=0.7.0",
    "pre-commit>=4.0.0",
    "ipython>=8.0.0",
]

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
